Add Kimi K3 TPU7x proxy workflow

+++ include/units/standard/VectorUnit.h
@@ -58,7 +58,7 @@ public:
 
 private:
     uint8_t idx;
-    int beats_per_wb;
+    uint64_t beats_per_wb;
   };
